don't fail if file doesn't exist

+++ makefile
@@ -56,7 +56,7 @@ start-db:
 	--name glamrous-postgres \
 	-p 5433:5432 \
 	glamrous-db
-	cp config.json config.json.bak
+	[[ -e config.json ]] && cp config.json config.json.bak
 	cp config.json.test config.json
 	DB_IP=sudo docker inspect --format '{{ .NetworkSettings.IPAddress }}' "glamrous-postgres"
 	sed -ie "s/IPADDRESS/${DB_IP}/g" config.json
@@ -81,4 +81,4 @@ clean:
 	rm -f .pylintrc
 
 reset-config:
-	cp config.json.bak config.json
+	[[ -e config.json.bak ]] && cp config.json.bak config.json
